fastmcp-simple/fastmcp_simple/agent/agent.py
from pathlib import Path

from google.adk.agents import Agent
from google.adk.tools.mcp_tool.mcp_toolset import MCPToolset
from mcp import StdioServerParameters

# IMPORTANT: Dynamically compute the absolute path to your server.py script
PATH_TO_MCP_SERVER_SCRIPT = str((Path(__file__).parent.parent / "server.py").resolve())

print(PATH_TO_MCP_SERVER_SCRIPT)


root_agent = Agent(
    name="greeting_agent",
    # https://ai.google.dev/gemini-api/docs/models
    model="gemini-2.5-flash-preview-04-17",
    description="Greeting agent",
    instruction="""
     You are a highly proactive and efficient assistant
    """,
    tools=[
        MCPToolset(
            connection_params=StdioServerParameters(
                command="python3",
                args=[PATH_TO_MCP_SERVER_SCRIPT],
            )
        )
    ],
)
